server/ahri/requirements.txt
django
pymongo
djangorestframework
django-cors-headers